working on dashboard state
