+++ frontend/debug.log
@@ -1,0 +1,8 @@
+[1215/152300.960:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/152300.961:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/152300.961:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/152300.961:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/152300.961:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/152300.962:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/152300.962:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/152300.962:ERROR:crash_report_database_win.cc(469)] failed to stat report
