+++ frontend/src/nav/dashboard/skills.js
@@ -1,18 +1,24 @@
-
 import useSignUpForm from './InputHooks';
-import React from 'react'
+import React from 'react';
 
-import { Form, Input, Button, Typography } from 'antd'
-const { Title } = Typography
+import { Form, Input, Button, Typography } from 'antd';
+const { Title } = Typography;
 
 const Skills = () => {
-  const { inputs, handleInputChange, handleSubmit } = useSignUpForm({
-    skills: '',
-  });
+  const handleOnclick = () => {
+    alert('Successfully saved!'); //it will send data to mongodb
+  };
+
+  const { inputs, handleInputChange, handleSubmit } = useSignUpForm(
+    {
+      skills: '',
+    },
+    handleOnclick,
+  );
 
   return (
     <div>
-    <Title level={3}>Skills</Title>
+      <Title level={3}>Skills</Title>
       <Form onSubmit={handleSubmit} autoComplete="off">
         <Form.Item label="Describe Your Skills">
           <Input
@@ -24,8 +30,8 @@ const Skills = () => {
           />
         </Form.Item>
         <Form.Item>
-          <Button type="primary submit">
-            <a href="../profile">Next</a>
+          <Button type="primary submit" onClick={handleSubmit}>
+            Submit
           </Button>
         </Form.Item>
       </Form>
