+++ frontend/src/nav/profile silinecek/skills.js
