+++ frontend/src/nav/profile.js
@@ -1,13 +1,7 @@
 import React from 'react';
 
-import ResumeDashboard from './resume/dashboard';
-
 const Profile = () => {
-  return (
-    <div>
-      <ResumeDashboard />
-    </div>
-  );
+  return <div>profile mapping</div>;
 };
 
 export default Profile;
